requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
JPype1>=1.2.0
jaydebeapi>=1.2.0
tkcalendar>=1.6.1
toml>=0.10.0
plotly>=5.15.0
streamlit-aggrid>=0.3.0
